pyproject.toml
[tool.poetry]
name = "foltz"
version = "0.1.0"
description = ""
authors = ["Pavel14701 <pavelkutia@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.3"
python-dotenv = "^1.0.1"
django-redis = "^5.4.0"
psycopg2 = "^2.9.10"
pillow = "^11.0.0"
requests = "^2.32.3"
redis = "^5.2.0"
gunicorn = "^23.0.0"
django-csp = "^3.8"
whitenoise = "^6.8.2"
pytils = "^0.4.1"
pytelegrambotapi = "^4.25.0"
djangorestframework = "^3.15.2"
django-compressor = "^4.5.1"
djangorestframework-simplejwt = "^5.3.1"
django-redis-sessions = "^0.6.2"
django-filter = "^24.3"
django-autocomplete-light = "^3.11.0"
graphene-django = "^3.2.2"
django-graphql-jwt = "^0.4.0"
jsonschema = "^4.23.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
